Temporary Island support.
Compiled but not tested. Don't know how to check it is run on island suppurt or not. 
It can be build and tested with/without "SNS_ISLAND_INCLUDE_AK0991X" option. But I don't know it works as designed. The ak0991x_dae_if_island.c is not implemented yet.